chore: Move genre queries
