+++ mobile/src/navigation/screens/genres/View.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from "react";
 import { useTranslation } from "react-i18next";
 
-import { useGenres } from "~/queries/genre";
+import { useGenres } from "~/data/genre/queries";
 import { useViewLayout } from "~/stores/ViewPreference/hooks/useViewLayout";
 import { useViewOrder } from "~/stores/ViewPreference/hooks/useViewOrder";
